Slow down ball when it bounces
Implemented a very basic form of friction, where the speed is reduced
by a fixed amount every time the ball bounces on a wall. Only the speed
perpendicular to the wall is affected, parallel speed is unchanged.
There is also a check for very low speed when bouncing. In that case,
the speed is set to zero and the position forced along the wall.